+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Changed
+- Relax `icanboogie/inflector` version dependency.
+
 ### Fixed
 - Fetching non-existent or dynamic properties no longer calls refresh().