+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.7 - 2018-01-21
 ### Added
 - Can set global locale instance **used by error instances with `Errors::setGlobalLocale()`.